feat(profile): add preview modal and refactor SupplierDetailPage (#93)
